Adding schemas to omml string

+++ index.js
@@ -6,18 +6,67 @@ var crypto = require('crypto');
 var Deferred = require('jsdeferred').Deferred;
 var imagemagick = require('imagemagick');
 var self = this;
+var transformedString = '';
 /**
  * Enables debugging messages. If true result file doesn't disappears
  * @const
  * @type {boolean}
  */
-var DEBUG = true;
+var DEBUG = false;
 /**
  * Array that contains availible file formats
  * @const
  * @type {Array.<string>}
  */
 var availibleFormats = ['jpeg', 'png', 'gif', 'jpg'];
+
+/**
+ * Strings with schemas needed for conversion
+ * @const
+ */
+var SCHEMAS_STRINGS = [{
+						name: 'xmlns:m',
+						definition: '"http://schemas.openxmlformats.org/officeDocument/2006/math"'
+					}, {
+						name: 'xmlns:mml',
+						definition: '"http://www.w3.org/1998/Math/MathML"'
+					}, {
+						name: 'xmlns:o',
+						definition: '"urn:schemas-microsoft-com:office:office"'
+					}, {
+						name: 'xmlns:r',
+						definition: '"http://schemas.openxmlformats.org/officeDocument/2006/relationships"'
+					}, {
+						name: 'xmlns:v',
+						definition: '"urn:schemas-microsoft-com:vml"'
+					}, {
+						name: 'xmlns:w',
+						definition: '"http://schemas.openxmlformats.org/wordprocessingml/2006/main"'
+					}, {
+						name: 'xmlns:w10',
+						definition: '"urn:schemas-microsoft-com:office:word"'
+					}, {
+						name: 'xmlns:wp',
+						definition: '"http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing"'
+					}];
+
+/**
+ * Adds missing schemas to omml string
+ *
+ * @returns {string} Result omml string with schemas
+ */
+function addSchemas(omml) {
+	var omath_index = omml.indexOf('>');
+	if(omath_index < 0)
+		return;
+	var result_omml = omml.substr(0, omath_index);
+	for(var i = 0; i < SCHEMAS_STRINGS.length; i++)
+		if(result_omml.indexOf(SCHEMAS_STRINGS[i].name)<0)
+			result_omml += ' '+SCHEMAS_STRINGS[i].name+'='+SCHEMAS_STRINGS[i].definition;
+	result_omml += omml.substring(omath_index, omml.length);
+	return result_omml;
+}
+
 
 /**
  *
@@ -49,7 +98,7 @@ function executeExternal() {
  * Read results of previous function
  * @see executeExternal
  *
- * @returns {{string Image data, { number width, number height} dimensions Contain image features }}
+ * @returns {{string Image data, { number width, number height} dimensions Contain image features, string file_path Path where image file saved }}
  *
  */
 function readResult(tmp_name) {
@@ -67,8 +116,10 @@ function readResult(tmp_name) {
 				width: features.width,
 				height: features.height
 			};
-			if (!DEBUG)
+			if (!DEBUG && self.options.remove_file)
 				fs.unlinkSync(tmp_name);
+			else
+				image.file_path = tmp_name;
 			d.call(image);
 		});
 	});
@@ -83,11 +134,11 @@ function readResult(tmp_name) {
  * @requires cb
  *
  * @param {string} omml String with Office MathML
- * @param {[{[encoding=utf8]: string In which encoding image data will be returned, [backgroundColor=white]: string, [fontColor=40]: number, [file_type=png]: string File type for image}]} options Must of them identical with mml2xxx generator
+ * @param {[{[encoding=utf8]: string In which encoding image data will be returned, [backgroundColor=white]: string, [fontColor=40]: number, [file_type=png]: string File type for image, [remove_file=true]: boolean Do you need to delete a file }]} options Must of them identical with mml2xxx generator
  * @param {function(string Contain image data in given encoding, Error)} cb Callback
  *
  */
-exports.renderFromString = function(omml, options, cb) {
+exports.renderFromString = function(omml, options, cb) {	
 	if (arguments.length == 2 && typeof arguments[1] == 'function' && typeof arguments[0] == 'string') {
 		cb = options;
 		options = null;
@@ -100,11 +151,16 @@ exports.renderFromString = function(omml, options, cb) {
 		this.options = {};
 	else
 		this.options = options;
+
+	omml = addSchemas(omml);
+
 	this.options.encoding = this.options.encoding || 'utf8';
 	this.options.backgroundColor = this.options.backgroundColor || 'white';
 	this.options.fontSize = this.options.fontSize || 40;
 	this.options.file_type = this.options.file_type || 'png';
 	this.options.file_type = this.options.file_type.replace('.', '').trim();
+	if(this.options.remove_file == null)
+		this.options.remove_file = true;
 	if (availibleFormats.indexOf(this.options.file_type) == -1)
 		cb(null, new Error('Wrong file_type passed to renderer'));
 	try {
@@ -122,4 +178,5 @@ exports.renderFromString = function(omml, options, cb) {
 		.error(function(error) {
 			cb(null, new Error('Error while converting: ' + error+'\n'+error.stack));
 		});
+	return;
 }
